<refactor> Modify code name to look intuitive(stopwatch_service -> stopwatch_timer_service)
