+++ lib/components/timer/timer_select_book_button.dart
@@ -6,7 +6,7 @@ import '../../constants/colors.dart';
 import '../../constants/text_styles.dart';
 import '../../services/pomodoro_timer_service.dart';
 import '../../services/reading_time_service.dart';
-import '../../services/stopwatch_service.dart';
+import '../../services/stopwatch_timer_service.dart';
 import 'timer_select_book_menu.dart';
 
 class BookSelectionWidget extends StatefulWidget {
